LPS-102450 Use getLong rather than getInt when referencing an Id column

+++ modules/apps/dynamic-data-mapping/dynamic-data-mapping-service/src/main/java/com/liferay/dynamic/data/mapping/internal/upgrade/v2_0_0/UpgradeDDMFormInstanceRecordVersion.java
@@ -83,7 +83,7 @@ public class UpgradeDDMFormInstanceRecordVersion extends UpgradeProcess {
 				ps2.setLong(12, rs.getLong("statusByUserId"));
 				ps2.setString(13, rs.getString("statusByUserName"));
 				ps2.setTimestamp(14, rs.getTimestamp("statusDate"));
-				ps2.setInt(15, rs.getInt("DDMStorageId"));
+				ps2.setLong(15, rs.getLong("DDMStorageId"));
 
 				deleteDDLRecordVersion(recordVersionId);
